requirements.txt
litestar
msgspec
requests
frida
pure-python-adb
pycryptodome